+++ src/exactly_lib/test_case_utils/expression/syntax_documentation.py
@@ -1,17 +1,17 @@
 import itertools
-from typing import List
+from typing import List, Sequence
 
 from exactly_lib.common.help.syntax_contents_structure import SyntaxElementDescription, InvokationVariant, \
     cli_argument_syntax_element_description
 from exactly_lib.definitions import formatting
 from exactly_lib.definitions.argument_rendering import cl_syntax
-from exactly_lib.definitions.entity.concepts import SYMBOL_CONCEPT_INFO
-from exactly_lib.definitions.entity.syntax_elements import SYMBOL_NAME_SYNTAX_ELEMENT
+from exactly_lib.definitions.entity import syntax_elements, concepts
 from exactly_lib.util.cli_syntax.elements import argument as a
 from exactly_lib.util.textformat.structure.core import ParagraphItem
 from exactly_lib.util.textformat.textformat_parser import TextParser
 from .grammar import Grammar, SimpleExpressionDescription, OperatorExpressionDescription
 from ...definitions.cross_ref.app_cross_ref import SeeAlsoTarget
+from ...definitions.entity.syntax_elements import SyntaxElementInfo
 
 
 def syntax_element_description(grammar: Grammar) -> SyntaxElementDescription:
@@ -25,7 +25,7 @@ class Syntax:
                                          self.grammar.concept.syntax_element)
 
         self._tp = TextParser({
-            'symbol_concept': formatting.concept_(SYMBOL_CONCEPT_INFO),
+            'symbol_concept': formatting.concept_name_with_formatting(concepts.SYMBOL_CONCEPT_INFO.name),
             'concept_name': self.grammar.concept.name,
         })
 
@@ -62,11 +62,26 @@ class Syntax:
         ]
 
     def invokation_variants_symbol_ref(self) -> List[InvokationVariant]:
-        symbol_argument = a.Single(a.Multiplicity.MANDATORY,
-                                   SYMBOL_NAME_SYNTAX_ELEMENT.argument)
-        iv = InvokationVariant(cl_syntax.cl_syntax_for_args([symbol_argument]),
-                               self._symbol_ref_description())
-        return [iv]
+        def invokation_variant(syntax_element: SyntaxElementInfo,
+                               description: Sequence[ParagraphItem]) -> InvokationVariant:
+            return InvokationVariant(cl_syntax.cl_syntax_for_args([
+                a.Single(a.Multiplicity.MANDATORY,
+                         syntax_element.argument)
+            ]),
+                description
+            )
+
+        description_of_sym_ref = self._symbol_ref_description()
+        description_of_sym_name = description_of_sym_ref + self._symbol_name_additional_description()
+
+        return [
+            invokation_variant(syntax_elements.SYMBOL_REFERENCE_SYNTAX_ELEMENT,
+                               description_of_sym_ref,
+                               ),
+            invokation_variant(syntax_elements.SYMBOL_NAME_SYNTAX_ELEMENT,
+                               description_of_sym_name,
+                               ),
+        ]
 
     def invokation_variants_complex(self) -> List[InvokationVariant]:
         def invokation_variant_of(operator_name: str,
@@ -124,20 +139,31 @@ class Syntax:
     def _symbol_ref_description(self) -> List[ParagraphItem]:
         return self._tp.fnap(_SYMBOL_REF_DESCRIPTION)
 
+    def _symbol_name_additional_description(self) -> List[ParagraphItem]:
+        return self._tp.fnap(_SYMBOL_NAME_ADDITIONAL_DESCRIPTION)
+
 
 def _see_also_targets_for_expr(expressions_dict: dict) -> iter:
     from_expressions = list(itertools.chain.from_iterable(
         map(lambda expr: expr.syntax.see_also_targets,
             expressions_dict.values())))
 
-    always = [SYMBOL_NAME_SYNTAX_ELEMENT.cross_reference_target]
+    always = [
+        syntax_elements.SYMBOL_NAME_SYNTAX_ELEMENT.cross_reference_target,
+        syntax_elements.SYMBOL_REFERENCE_SYNTAX_ELEMENT.cross_reference_target,
+    ]
 
     return from_expressions + always
 
 
 _SYMBOL_REF_DESCRIPTION = """\
-Reference to a {symbol_concept},
+Reference to {symbol_concept:a},
 that must have been defined as {concept_name:a}.
+"""
+
+_SYMBOL_NAME_ADDITIONAL_DESCRIPTION = """\
+A string that is not the name of {concept_name:a}
+is interpreted as the name of {symbol_concept:a}.
 """
 
 _GLOBAL_DESCRIPTION = """\