+++ test/exactly_lib_test/instructions/assert_/contents_of_dir/common.py
@@ -3,13 +3,14 @@ import unittest
 from exactly_lib.instructions.assert_ import contents_of_dir as sut
 from exactly_lib.section_document.element_parsers.instruction_parser_exceptions import \
     SingleInstructionInvalidArgumentException
-from exactly_lib.symbol.symbol_syntax import symbol_reference_syntax_for_name
 from exactly_lib.util.logic_types import ExpectationType
 from exactly_lib.util.name_and_value import NameAndValue
 from exactly_lib_test.common.help.test_resources.check_documentation import suite_for_instruction_documentation
 from exactly_lib_test.instructions.assert_.contents_of_dir.test_resources import instruction_arguments as args
 from exactly_lib_test.instructions.assert_.contents_of_dir.test_resources import tr
 from exactly_lib_test.section_document.test_resources.misc import ARBITRARY_FS_LOCATION_INFO
+from exactly_lib_test.symbol.test_resources.symbol_syntax import \
+    NOT_A_VALID_SYMBOL_NAME_NOR_PRIMITIVE_GRAMMAR_ELEMENT_NAME
 from exactly_lib_test.test_case_utils.files_matcher.test_resources.arguments_building import \
     AssertionVariantArgumentsConstructor
 from exactly_lib_test.test_case_utils.parse.test_resources.single_line_source_instruction_utils import \
@@ -47,7 +48,9 @@ class TestParseInvalidSyntax(tr.TestCaseBaseForParser):
                 'valid file argument, invalid check',
                 args.complete_arguments_constructor(
                     'file-name',
-                    InvalidAssertionVariantArgumentsConstructor(symbol_reference_syntax_for_name('invalidCheck'))
+                    InvalidAssertionVariantArgumentsConstructor(
+                        NOT_A_VALID_SYMBOL_NAME_NOR_PRIMITIVE_GRAMMAR_ELEMENT_NAME
+                    )
                 ),
             ),
         ]