+++ test/exactly_lib_test/symbol/test_resources/symbol_syntax.py
@@ -1,3 +1,5 @@
 NOT_A_VALID_SYMBOL_NAME = 'not/a/valid/symbol/name'
 
+NOT_A_VALID_SYMBOL_NAME_NOR_PRIMITIVE_GRAMMAR_ELEMENT_NAME = 'not/a/valid/symbol/name/nor-primitive/grammar/element'
+
 A_VALID_SYMBOL_NAME = 'a_valid_symbol_name'