+++ test/exactly_lib_test/symbol/z_package_suite.py
@@ -1,6 +1,6 @@
 import unittest
 
-from exactly_lib_test.symbol import restriction, data
+from exactly_lib_test.symbol import restriction, data, symbol_syntax
 from exactly_lib_test.symbol.data import z_package_suite as data
 from exactly_lib_test.symbol.err_msg import z_package_suite as err_msg
 from exactly_lib_test.symbol.logic import z_package_suite as logic
@@ -10,6 +10,7 @@ from exactly_lib_test.symbol.test_resources_test import z_package_suite as test_
 def suite() -> unittest.TestSuite:
     return unittest.TestSuite([
         test_resources_test.suite(),
+        symbol_syntax.suite(),
         err_msg.suite(),
         restriction.suite(),
         data.suite(),